services/sync_dal/requirements.txt
psycopg2-binary
gunicorn
flask
redis